requirements.txt
google-genai>=1.0.0
pystray>=0.19.5
Pillow>=10.0.0
keyboard>=0.13.5
pyperclip>=1.9.0
pywin32>=306
pyinstaller>=6.0.0
pytest>=8.0.0
